feat: flexible buffer & bitmap

+++ STM32N6570DK_LL/STM32N6570DK/FSBL/Src/Drivers/Simple/Src/simple_lcd_framebuffer.c
@@ -7,7 +7,7 @@
 
 #include "simple_lcd_framebuffer.h"
 
-volatile uint32_t lcd_bg_buffer[LCD_BG_WIDTH * LCD_BG_HEIGHT] __attribute__((section(".psram_bss"), aligned(32)));
+volatile uint16_t lcd_bg_buffer[LCD_BG_WIDTH * LCD_BG_HEIGHT] __attribute__((section(".psram_bss"), aligned(32)));
 volatile uint32_t lcd_fg_buffer[LCD_FG_WIDTH * LCD_FG_HEIGHT] __attribute__((section(".psram_bss"), aligned(32)));
 
 //void LCD_Fill(uint32_t color){
